Upgrade/skill system

+++ Assets/Scriptable/Script/Skill/SkillState.cs
@@ -1,9 +1,10 @@
+using TMPro;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "New SkillState", menuName = "Scriptable/SkillState")]
 public class SkillState : ScriptableObject
 {
     public string nameSkill;
-    public float price;
+    public int price;
     public Sprite sprite;
 }
